src/test/java/com/self/ddyoung/rice/controller/UserControllerTest.java
package com.self.ddyoung.rice.controller;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.mock.web.MockServletContext;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.web.WebAppConfiguration;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import static org.hamcrest.Matchers.is;
import static org.hamcrest.Matchers.notNullValue;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

/**
 * Created by sanbian on 2017/8/8.
 */
//这是JUnit的注解，通过这个注解让SpringJUnit4ClassRunner这个类提供Spring测试上下文。
@RunWith(SpringJUnit4ClassRunner.class)
//这是Spring Boot注解，为了进行集成测试，需要通过这个注解加载和配置Spring应用上下
@SpringBootTest(classes = MockServletContext.class)
@WebAppConfiguration
public class UserControllerTest {


    private MockMvc mockMvc;

    @Before
    public void setUp() throws Exception {
        mockMvc = MockMvcBuilders.standaloneSetup(new UserController()).build();
    }



    @Test
    public void testGetUser() throws Exception {
//        UserDO userDO = new UserDO();
//        userDO.setCreateTime(new Date());
//        userDO.setModifyTime(new Date());
//        userDO.setUserName("1111");
//        userDO.setPassword("1111");
//        userDO.setEmail("zdan68@163.com");
//        userDO.setNickName("nick1111");


        mockMvc.perform(get("/user/getUser?id=1"))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON_UTF8))
                .andExpect(jsonPath("$.errcode", is(0)))
                .andExpect(jsonPath("$.data", notNullValue()))
                .andExpect(jsonPath("$.data.id", is(1)));
    }

    @Test
    public void testAddUser() throws Exception {

    }
}